Changed map drawing system to heightmap stuff that was in the visualization lecture this morning Changed the map to better test the new drawing system Moved more stuff to the settings class for external save/loading Made the buttons on the menu do stuff, heh

+++ GreeningOfMars/src/Settings.cpp
@@ -29,7 +29,7 @@ void Settings::Save(char* filepath)
 	IniHandler::SetFileName(".\\settings.ini");
 	IniHandler::WriteInt("View", "Width", View::Width);
 	IniHandler::WriteInt("View", "Height", View::Height);
-	IniHandler::WriteFloat("View", "FOV", View::FOV);
+	IniHandler::WriteFloat("View", "FOV", (float)View::FOV);
 	IniHandler::WriteInt("View", "Fullscreen", View::Fullscreen);
 
 	IniHandler::WriteFloat("Mouse", "MovementRatio", Mouse::MovementRatio);
